Wolverine durability added
The documentation regarding durable messaging with EF Core & modular monolith was outdated which lead me to believe it is not possible. Luckily, I came across a blogpost by the maintainer that it is actually possible so now it is implemented.

Also added ValueGeneratedOnAdd to BlogPostId in BlogDbContext. With the HasDefaultValueSql the value was just an empty guid.